% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer_von_mises_test.R
\name{cramer_von_mises_tstat}
\alias{cramer_von_mises_tstat}
\title{This function calculates the realization of the cramer-von mises type test
on two given samples}
\usage{
cramer_von_mises_tstat(
  sample1,
  sample2,
  domain = c(0, 1),
  type = "fourier",
  n_basis = NULL,
  grid = NULL,
  mu_measure,
  n_func
)
}
\arguments{
\item{sample1:}{first sample, specified as a list where each element is one observation}

\item{sample2:}{second sample, specified as a list where each element is one observation}

\item{domain:}{Domain of the functional observations}

\item{type:}{Method used for the calculation of the test statistic tau;
possible values: 'fourier', 'grid'}

\item{n_basis:}{if basis based method is chosen in type, this determines the
number of basis functions used for the calculation}

\item{grid:}{if type grid is chosen, this specifies the grid for the function
comparison}

\item{mu_measure:}{function used for the generation of realizations of the random
function corresponding to the measure mu}

\item{n_func:}{number of functions used for the Monte-Carlo Integration in the
approximation of tau}
}
\value{
: The value of the t-statistic for the Cramer-von Mises test
}
\description{
This function calculates the realization of the cramer-von mises type test
on two given samples
}
