% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer_von_mises_test.R
\name{cramer_von_mises_tstat_grid}
\alias{cramer_von_mises_tstat_grid}
\title{This function calculates the realization of the cramer-von mises type test
on two given samples (for grid based variant)}
\usage{
cramer_von_mises_tstat_grid(
  sample1,
  sample2,
  domain = c(0, 1),
  basis,
  w_func,
  rho,
  u_sample_func,
  grid = NULL,
  n_func,
  ...
)
}
\arguments{
\item{sample1:}{first sample, specified as a list where each element is one observation}

\item{sample2:}{second sample, specified as a list where each element is one observation}

\item{domain:}{Domain of the functional observations}

\item{basis:}{basis generated by fda}

\item{w_func:}{weight function for construction of measure}

\item{rho:}{a vector of length n_basis from a square summable sequence that is used
as scaling terms for the error terms}

\item{u_sample_func:}{function that can be used to sample from for the error terms
around the specified mean}

\item{n_basis:}{if basis based method is chosen in type, this determines the
number of basis functions used for the calculation}

\item{grid:}{if type grid is chosen, this specifies the grid for the function
comparison}

\item{n_func:}{number of functions used for the Monte-Carlo Integration in the
approximation of tau}

\item{...:}{additional arguments given to u_sample_func}
}
\value{
: The value of the t-statistic for the Cramer-von Mises test
}
\description{
This function calculates the realization of the cramer-von mises type test
on two given samples (for grid based variant)
}
