% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_perm_test_functions.R
\name{crit_val}
\alias{crit_val}
\title{This function returns a critical value for a chosen significance level
given a vector of realizations of the test statistic}
\usage{
crit_val(realizations, alpha = 0.05)
}
\arguments{
\item{realizations:}{vector of realizations of the test statistic}

\item{alpha:}{significance level of the test}
}
\value{
: Critical Value for the test
}
\description{
This function returns a critical value for a chosen significance level
given a vector of realizations of the test statistic
}
