% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_distribution_function.R
\name{empirical_dist_func}
\alias{empirical_dist_func}
\title{Returns the value of an empirical distribution function of a given sample
for a given function}
\usage{
empirical_dist_func(
  sample,
  func,
  domain = c(0, 1),
  grid = NULL,
  fourier = FALSE,
  n_fourier_basis = NULL
)
}
\arguments{
\item{sample:}{Sample of functional data. This has to be in a slighly weird
format as the sample has to be provided as a list of singular fd objects,
each representing a single observation. A wrapper function to transform from
the typical format to this is provided in:}

\item{func:}{Function to evaluate the empirical distribution function at}

\item{domain:}{Domain of the functions in both the sample and the comparison
function}

\item{grid:}{grid used to decide whether the functions are always bigger}

\item{fourier:}{Boolean that decided whether the method for fourier zeroes is
used to compare the functions}

\item{n_fourier_basis:}{if fourier == TRUE, this decides the number of
fourier basis functions to use}
}
\value{
An approximated value of the empirical distribution function
}
\description{
Returns the value of an empirical distribution function of a given sample
for a given function
}
