% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_distribution_function_fourier.R
\name{empirical_dist_func_fourier}
\alias{empirical_dist_func_fourier}
\title{Returns the value of an empirical distribution function of a given sample
for a given function}
\usage{
empirical_dist_func_fourier(sample, func, domain = c(0, 1), n_fourier_basis)
}
\arguments{
\item{sample:}{Sample of functional data}

\item{func:}{Function to evaluate the empirical distribution function at}

\item{domain:}{Domain of the functions in both the sample and the comparison
function}

\item{n_fourier_basis:}{if fourier == TRUE, this decides the number of
fourier basis functions to use}
}
\value{
An approximated value of the empirical distribution function
}
\description{
Returns the value of an empirical distribution function of a given sample
for a given function
}
