% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_distribution_function.R
\name{empirical_dist_func_grid}
\alias{empirical_dist_func_grid}
\title{Returns the value of an empirical distribution function of a given sample
for a given function}
\usage{
empirical_dist_func_grid(sample, func, domain = c(0, 1), grid = NULL)
}
\arguments{
\item{sample:}{Sample of functional data}

\item{func:}{Function to evaluate the empirical distribution function at}

\item{domain:}{Domain of the functions in both the sample and the comparison
function}

\item{grid:}{grid used to decide whether the functions are always bigger}
}
\value{
An approximated value of the empirical distribution function
}
\description{
Returns the value of an empirical distribution function of a given sample
for a given function
}
