% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_construct.R
\name{fourier_basis_coef_means}
\alias{fourier_basis_coef_means}
\title{This function calculates the means of the fourier coefficients as described
in section 5 of Bugni, Horowitz (2021) for a chosen basis and a function  w}
\usage{
fourier_basis_coef_means(w_func, basis = "fourier", n_basis, domain = c(0, 1))
}
\arguments{
\item{w_func:}{function that is large in the areas of the domain the
investigator expects differences in the distribution functions}

\item{basis:}{type of functional basis to use (has to be an orthogonal basis)}

\item{n_basis:}{number of basis functions}

\item{domain:}{vector of two points (start and endpoint of the closed interval)}
}
\value{
: Vector of length n_basis containing the resulting means of the
fourier coefficients
}
\description{
This function calculates the means of the fourier coefficients as described
in section 5 of Bugni, Horowitz (2021) for a chosen basis and a function  w
}
