% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_construct_eigen.R
\name{fourier_coef_mean_eigen}
\alias{fourier_coef_mean_eigen}
\title{This function calculates the desired mean of a fourier coefficient for
a single basis function}
\usage{
fourier_coef_mean_eigen(w_func, basis_func, domain = c(0, 1))
}
\arguments{
\item{w_func:}{function that is large in the areas of the domain the
investigator expects differences in the distribution functions}

\item{basis_func:}{basis function to calculate the fourier coefficient for}

\item{domain:}{vector of two points (start and endpoint of the closed interval)}
}
\value{
: Double that is the desired mean of the fourier coefficient
}
\description{
This function calculates the desired mean of a fourier coefficient for
a single basis function
}
