% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_construct.R
\name{fourier_coef_sample}
\alias{fourier_coef_sample}
\title{This function samples a desired distribution of fourier coefficients
around a mean that's determined by the function w}
\usage{
fourier_coef_sample(
  n = 1,
  w_func,
  basis = "fourier",
  n_basis,
  u_sample_func,
  domain = c(0, 1),
  ...
)
}
\arguments{
\item{w_func:}{function that is large in the areas of the domain the
investigator expects differences in the distribution functions}

\item{basis:}{type of functional basis to use (has to be an orthogonal basis)}

\item{n_basis:}{number of basis functions}

\item{domain:}{vector of two points (start and endpoint of the closed interval)}

\item{u_sample_func:}{function that can be used to sample from for the error terms
around the specified mean - each observation should be a vector of length n_sample
and the function should have an argument n to determine the number of observations sampled
(for example multivariate normal of appropriate dimension)}

\item{...:}{further parameters that are given to u_sample_func}
}
\value{
: A list of length n containing vectors of sampled fourier coefficients
of length n_basis
}
\description{
This function samples a desired distribution of fourier coefficients
around a mean that's determined by the function w
}
