% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_basis.R
\name{fourier_eval_matrix}
\alias{fourier_eval_matrix}
\title{This is a helper function that evaluates a functional object given as its
fourier coefficients at points}
\usage{
fourier_eval_matrix(x, coefs, domain = c(0, 1))
}
\arguments{
\item{x:}{point or vector of points to evaluate the function at}

\item{ceofs:}{matrix of coefficients identifying the functional objects}

\item{domain:}{domain of the basis}

\item{num_func:}{Which fourier function to evaluate}
}
\description{
This is a helper function that evaluates a functional object given as its
fourier coefficients at points
}
