% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_basis.R
\name{fourier_func}
\alias{fourier_func}
\title{This is a helper function to not use the fda package in the function evaluation and comparison}
\usage{
fourier_func(x, domain = c(0, 1), num_func)
}
\arguments{
\item{x:}{point or vector of points to evaluate the function at}

\item{domain:}{domain of the basis}

\item{num_func:}{Which fourier function to evaluate}
}
\description{
This is a helper function to not use the fda package in the function evaluation and comparison
}
