% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_estimators.R
\name{fourier_mean_estimator}
\alias{fourier_mean_estimator}
\title{This function finds the estimated mean function from a given sample using
an approximation using a truncated Fourier basis}
\usage{
fourier_mean_estimator(sample, domain, n_basis, grid)
}
\arguments{
\item{sample:}{sample, specified as a list where each element is one observation}

\item{domain:}{vector with beginning and endpoint of the closed interval
that is the domain of the stochastic processes}

\item{n_basis:}{the number of basis functions used in the approximation}

\item{grid:}{grid used for the approximation of the mean function}
}
\value{
Placeholder
}
\description{
This function finds the estimated mean function from a given sample using
an approximation using a truncated Fourier basis
}
