% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_comparison.R
\name{fourier_zeroes}
\alias{fourier_zeroes}
\title{This function calculates the zeroes of the difference of two functions
given in terms of the fourier coefficients of the same fourier basis}
\usage{
fourier_zeroes(func_a, func_b, domain = c(0, 1))
}
\arguments{
\item{func_a:}{Fourier coefficients for Function one}

\item{func_b:}{Fourier coefficients for Function two}

\item{domain:}{interval to check}
}
\value{
The zeroes of the difference function
}
\description{
This function calculates the zeroes of the difference of two functions
given in terms of the fourier coefficients of the same fourier basis
}
