% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_distribution_function.R
\name{func_comparison}
\alias{func_comparison}
\title{This function checks if one function is bigger than another function for all
points in a closed interval}
\usage{
func_comparison(func_a, func_b, domain = c(0, 1))
}
\arguments{
\item{func_a:}{Function one}

\item{func_b:}{Function two}

\item{domain:}{interval to check}
}
\value{
TRUE or FALSE depending on wheter func_a is always bigger than
func_b
}
\description{
This function checks if one function is bigger than another function for all
points in a closed interval
}
