% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_comparison.R
\name{func_comparison_fourier}
\alias{func_comparison_fourier}
\title{This function checks if one function is bigger than another function for all
points in a closed interval}
\usage{
func_comparison_fourier(func_a, func_b, domain = c(0, 1))
}
\arguments{
\item{func_a:}{fd object for Function one}

\item{func_b:}{fd object for Function two}

\item{domain:}{interval to check}
}
\value{
TRUE or FALSE depending on wheter func_a is always bigger than
func_b
}
\description{
This function checks if one function is bigger than another function for all
points in a closed interval
}
