% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_comparison.R
\name{func_comparison_grid}
\alias{func_comparison_grid}
\title{This function checks if one function is bigger than another function for all
points in a closed interval}
\usage{
func_comparison_grid(func_a, func_b, domain = c(0, 1), grid)
}
\arguments{
\item{func_a:}{Fourier coefficients for Function one}

\item{func_b:}{Fourier coefficients for Function two}

\item{domain:}{interval to check}

\item{grid:}{grid of points used to checks}
}
\value{
TRUE or FALSE depending on wheter func_a is always bigger than
func_b
}
\description{
This function checks if one function is bigger than another function for all
points in a closed interval
}
