% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_sample_transform.R
\name{func_sample_transform}
\alias{func_sample_transform}
\title{This function transforms a sample of functional data as it is typically generated
by functions from the package fda into the format that is used for example
for the function empirical_dist_func}
\usage{
func_sample_transform()
}
\arguments{
\item{original_sample:}{}
}
\description{
This function transforms a sample of functional data as it is typically generated
by functions from the package fda into the format that is used for example
for the function empirical_dist_func
}
