% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{func_sample_transform2}
\alias{func_sample_transform2}
\title{This function transforms a sample of functional data as it is used by this package
into the format that is used by the package fda}
\usage{
func_sample_transform2(orgnl_sample)
}
\arguments{
\item{orgnl_sample:}{fda object representing multiple observations;
as a list of individual fda objects}
}
\value{
: sample in the format that is used by fda
}
\description{
This function transforms a sample of functional data as it is used by this package
into the format that is used by the package fda
}
