% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_construct.R
\name{function_sample_eigen}
\alias{function_sample_eigen}
\title{This function uses fourier coefficients generated by the function fourier_coef_sample
to generate realizations of the random function associated with the
measure mu. It uses the empirical eigenfunctions as the basis for the generation}
\usage{
function_sample_eigen(
  w_func,
  eigen_func_obj,
  n_basis,
  rho,
  u_sample_func,
  domain = c(0, 1),
  ...
)
}
\arguments{
\item{w_func:}{function that is large in the areas of the domain the
investigator expects differences in the distribution functions}

\item{eigen_func_obj:}{fda object containing the empirical eigenfunctions}

\item{n_basis:}{number of basis functions (here: empirical eigenfunctions)}

\item{rho:}{a vector of length n_basis from a square summable sequence that is used
as scaling terms for the error terms}

\item{domain:}{vector of two points (start and endpoint of the closed interval)}

\item{u_sample_func:}{function that can be used to sample from for the error terms
around the specified mean}

\item{...:}{further parameters that are given to u_sample_func}
}
\value{
: A functional object that represents the sampled function
}
\description{
This function uses fourier coefficients generated by the function fourier_coef_sample
to generate realizations of the random function associated with the
measure mu. It uses the empirical eigenfunctions as the basis for the generation
}
