% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_estimators.R
\name{linear_interpolation}
\alias{linear_interpolation}
\title{This function approximates an observation on a specified grid
using linear interpolation}
\usage{
linear_interpolation(observation, grid)
}
\arguments{
\item{observation:}{Placeholder}

\item{grid:}{Placeholder}
}
\value{
Placeholder
}
\description{
This function approximates an observation on a specified grid
using linear interpolation
}
