% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_estimators.R
\name{mean_estimator}
\alias{mean_estimator}
\title{This function finds the estimated mean function from a given sample}
\usage{
mean_estimator(
  sample,
  interpolation_mode = "linear",
  domain = c(0, 1),
  n_basis = NA,
  grid
)
}
\arguments{
\item{sample:}{sample, specified as a list where each element is one observation}

\item{interpolation_mode:}{string that determines the mode of interpolation between
discrete measurement points. Options: 'linear', 'fourier', 'bspline'}

\item{domain:}{vector with beginning and endpoint of the closed interval
that is the domain of the stochastic processes}

\item{n_basis:}{if interpolation mode is a basis based method,
this determines the number of basis functions used in the approximation}

\item{grid:}{grid used for the approximation of the mean function}
}
\value{
Placeholder
}
\description{
This function finds the estimated mean function from a given sample
}
