% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means_test.R
\name{means_crit_value}
\alias{means_crit_value}
\title{This function calculates / approximates the critical value for the means based test}
\usage{
means_crit_value(
  alpha = 0.05,
  full = TRUE,
  approxQ,
  sample1,
  sample2,
  interpolation_mode = "linear",
  domain = c(0, 1),
  n_basis = NULL,
  grid
)
}
\arguments{
\item{sample1:}{first sample, specified as a list where each element is one observation}

\item{sample2:}{second sample, specified as a list where each element is one observation}

\item{interpolation_mode:}{string that determines the mode of interpolation between
discrete measurement points. (Recommendation: 'bspline')}

\item{domain:}{vector with beginning and endpoint of the closed interval
that is the domain of the stochastic processes}

\item{n_basis:}{if interpolation mode is choses as a basis based method,
this determines the number of basis functions used in the approximation}

\item{grid:}{grid used for the approximation of the mean function}
}
\value{
The value of the t statistic for means based test
for the chosen samples.
}
\description{
This function calculates / approximates the critical value for the means based test
}
