% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means_test.R
\name{means_test}
\alias{means_test}
\title{This function performs the mean-based test for two samples with a chosen
set of parameters.}
\usage{
means_test(
  sample1,
  sample2,
  interpolation_mode = "linear",
  domain = c(0, 1),
  n_basis = NA,
  grid
)
}
\arguments{
\item{sample1:}{first sample, specified as a list where each element is one observation}

\item{sample2:}{second sample, specified as a list where each element is one observation}

\item{interpolation_mode:}{string that determines the mode of interpolation between
discrete measurement points.}

\item{domain:}{vector with beginning and endpoint of the closed interval
that is the domain of the stochastic processes}

\item{n_basis:}{if interpolation mode is choses as a basis based method,
this determines the number of basis functions used in the approximation}

\item{grid:}{grid used for the approximation of the mean function}

\item{...:}{additional arguments that are relayed to t_stat_func}
}
\value{
Placeholder
}
\description{
This function performs the mean-based test for two samples with a chosen
set of parameters.
}
