% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_test_variants.R
\name{means_tstats_par}
\alias{means_tstats_par}
\title{This function is aimed at the application for my master's thesis and is
experimental. It performs the approximation means based test parallelized
over the permutations.}
\usage{
means_tstats_par(
  cl,
  full = TRUE,
  approxQ = NULL,
  sample1,
  sample2,
  interpolation_mode = "linear",
  domain = c(0, 1),
  n_basis = NULL,
  grid = NULL
)
}
\arguments{
\item{cl:}{cluster object created by parallel package}

\item{full:}{Boolean that decides whether all combinations are considered
if FALSE, approxQ has to be provided}

\item{approxQ:}{integer specifying the number of combinations to be used for
the approximation of the critical value}

\item{sample1:}{first sample, specified as a list where each element is one observation}

\item{sample2:}{second sample, specified as a list where each element is one observation}

\item{interpolation_mode:}{string that determines the mode of interpolation between
discrete measurement points. (Recommendation: 'linear')}

\item{domain:}{vector with beginning and endpoint of the closed interval
that is the domain of the stochastic processes}

\item{n_basis:}{if interpolation mode is choses as a basis based method,
this determines the number of basis functions used in the approximation}

\item{grid:}{grid used for the approximation of the mean function}
}
\value{
Realizations of the t-statistic for the Mean based test
}
\description{
This function is aimed at the application for my master's thesis and is
experimental. It performs the approximation means based test parallelized
over the permutations.
}
