% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_functional_integration.R
\name{monte_carlo_integration}
\alias{monte_carlo_integration}
\title{This function performs Monte-Carlo integration over L2 for the test function
and two original samples for a given number of evaluation points (this is meant
for the Cramer-von Mises test statistic)}
\usage{
monte_carlo_integration(t_stat_func, n_points, sample1, sample2, ...)
}
\arguments{
\item{t_stat_func:}{function that calculates the test statistic on a given sample}

\item{n_points:}{number of random points used for the approximation}

\item{sample1:}{first sample, specified as a list where each element is one observation}

\item{sample2:}{second sample, specified as a list where each element is one observation}

\item{...:}{additional arguments for the function t_stat_func}
}
\value{
:
}
\description{
This function performs Monte-Carlo integration over L2 for the test function
and two original samples for a given number of evaluation points (this is meant
for the Cramer-von Mises test statistic)
}
