% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{next_cbn}
\alias{next_cbn}
\title{Get the next combination of size k from 1:n given cbn as an argument
describing the last combination. (Thanks to user Quantibex on stackoverflow
https://stackoverflow.com/users/2513407/quantibex)}
\usage{
next_cbn(cbn = NULL, n, k)
}
\arguments{
\item{cbn}{The last combination. If Null: return the first combination.}

\item{n}{The size of the full data set}

\item{k}{The size of sample 1}
}
\value{
The next combination.
}
\description{
Get the next combination of size k from 1:n given cbn as an argument
describing the last combination. (Thanks to user Quantibex on stackoverflow
https://stackoverflow.com/users/2513407/quantibex)
}
\examples{
next_cbn(cbn = c(1, 2, 3), n = 6, k = 3)
}
