% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu_test.R
\name{nu_realizations}
\alias{nu_realizations}
\title{This function calculates realizations of the cramer-von mises type test
on two given samples for a specified number of permutations}
\usage{
nu_realizations(sample1, sample2, full = FALSE, approxQ, domain)
}
\arguments{
\item{sample1:}{first sample, specified as a list where each element is one observation}

\item{sample2:}{second sample, specified as a list where each element is one observation}

\item{full:}{Boolean that decides whether all combinations are considered
if FALSE, approxQ has to be provided}

\item{approxQ:}{integer specifying the number of combinations to be used for
the approximation of the critical value}

\item{domain:}{Domain of the functional observations}
}
\value{
: Realizations of the t-statistic for the Cramer-von Mises test
}
\description{
This function calculates realizations of the cramer-von mises type test
on two given samples for a specified number of permutations
}
