% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{obs_generator_1}
\alias{obs_generator_1}
\title{This function generates an observation for testing purposes}
\usage{
obs_generator_1(domain = c(0, 1), n_points = 100)
}
\arguments{
\item{domain:}{beginning and endpoint of domain (closed interval)}

\item{n_points:}{number of measurement points in observation}
}
\value{
: An observation in the typical format
}
\description{
This function generates an observation for testing purposes
}
