% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test.R
\name{perm_crit_value}
\alias{perm_crit_value}
\title{This function derives / approximates the critical value for the
permutation test}
\usage{
perm_crit_value(alpha = 0.05, full = TRUE, approxQ, sample1, sample2)
}
\arguments{
\item{alpha:}{Size of the test}

\item{full:}{Boolean that decides whether all combinations are considered
if FALSE, approxQ has to be provided}

\item{approxQ:}{integer specifying the number of combinations to be used for
the approximation of the critical value}

\item{sample1:}{first sample, specified as a list where each element is one observation}

\item{sample2:}{second sample, specified as a list where each element is one observation}
}
\description{
This function derives / approximates the critical value for the
permutation test
}
