% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_procedure.R
\name{perm_tstats}
\alias{perm_tstats}
\title{This function calculates realizations of the test statistic over a specified number
of permutations of the data}
\usage{
perm_tstats(full = TRUE, approxQ = NULL, sample1, sample2, t_stat_func, ...)
}
\arguments{
\item{full:}{Boolean that decides whether all combinations are considered
if FALSE, approxQ has to be provided}

\item{approxQ:}{integer specifying the number of combinations to be used for
the approximation of the critical value}

\item{sample1:}{first sample, specified as a list where each element is one observation}

\item{sample2:}{second sample, specified as a list where each element is one observation}

\item{t_stat_func:}{function that calculates the test statistic on a given sample}

\item{...:}{additional arguments that are relayed to t_stat_func}
}
\description{
This function calculates realizations of the test statistic over a specified number
of permutations of the data
}
