% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_perm_test_functions.R
\name{perm_tstats_approx}
\alias{perm_tstats_approx}
\title{This function calculates realizations of the test statistic by considering
a specified number of combinations. The specific combinations that are considered
are chosen randomly with replacement.}
\usage{
perm_tstats_approx(approxQ, sample1, sample2, t_stat_func, ...)
}
\arguments{
\item{approxQ:}{integer specifying the number of combinations to be used for
the approximation of the critical value}

\item{sample1:}{first sample, specified as a list where each element is one observation}

\item{sample2:}{second sample, specified as a list where each element is one observation}

\item{t_stat_func:}{function that calculates the test statistic on a given sample}

\item{...:}{additional arguments that are relayed to t_stat_func}
}
\description{
This function calculates realizations of the test statistic by considering
a specified number of combinations. The specific combinations that are considered
are chosen randomly with replacement.
}
