% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_procedure.R
\name{perm_tstats_full}
\alias{perm_tstats_full}
\title{This function calculates all realizations of the test statistic by considering
all combinations of the data of the specified sample sizes}
\usage{
perm_tstats_full(Q, sample1, sample2, t_stat_func, ...)
}
\arguments{
\item{Q:}{possible number of combinations}

\item{sample1:}{first sample, specified as a list where each element is one observation}

\item{sample2:}{second sample, specified as a list where each element is one observation}

\item{t_stat_func:}{function that calculates the test statistic on a given sample}

\item{...:}{additional arguments that are relayed to t_stat_func}
}
\description{
This function calculates all realizations of the test statistic by considering
all combinations of the data of the specified sample sizes
}
