% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{quick_funcify}
\alias{quick_funcify}
\title{This function translates a sample generated with the sample_generator
functions into a functional data set}
\usage{
quick_funcify(sample, domain = c(0, 1), n_fourier_basis)
}
\arguments{
\item{sample:}{Sample to translate}

\item{domain:}{beginning and endpoint of domain (closed interval)}

\item{n_fourier_basis:}{number of fourier basis functions to use}
}
\value{
: A sample in the typical format
}
\description{
This function translates a sample generated with the sample_generator
functions into a functional data set
}
