% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_construct.R
\name{rho_construct}
\alias{rho_construct}
\title{This function implements the method of choosing the sequence of rhos
described in my thesis}
\usage{
rho_construct(sample, factor)
}
\arguments{
\item{sample:}{sample as a matrix of fourier coefficients}

\item{factor:}{factor to multiplt the empirical standard deviation with}
}
\description{
This function implements the method of choosing the sequence of rhos
described in my thesis
}
