% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_construct.R
\name{rho_draw}
\alias{rho_draw}
\title{This function is used to draw the rho coefficients used in the construction
of the mu measure as described in the paper. These have to falls into l2
However, it is often advisable to specify these yourself.}
\usage{
rho_draw(n_basis, coef_vector = NULL)
}
\description{
This function is used to draw the rho coefficients used in the construction
of the mu measure as described in the paper. These have to falls into l2
However, it is often advisable to specify these yourself.
}
