% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{sample_generator_1}
\alias{sample_generator_1}
\title{This function generates a sample for testing purposes}
\usage{
sample_generator_1(n_obs = 100, domain = c(0, 1))
}
\arguments{
\item{n_obs:}{Number of observations in sample}

\item{domain:}{beginning and endpoint of domain (closed interval)}
}
\value{
: A sample in the typical format
}
\description{
This function generates a sample for testing purposes
}
