% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{sample_inds}
\alias{sample_inds}
\title{Given an integer n and a vector smpl1 containing elements of 1:n,
return a list containing smpl1 and its counterpart smpl2}
\usage{
sample_inds(n, smpl1)
}
\arguments{
\item{n}{An integer specifying the sample 1:n to be split}

\item{smpl1}{a vector containing elements of 1:n}
}
\value{
A list containing two vectors smpl1 and smpl2. These are disjunct
and their union is 1:n.
}
\description{
Given an integer n and a vector smpl1 containing elements of 1:n,
return a list containing smpl1 and its counterpart smpl2
}
\examples{
sample_inds(n = 10, smpl1 = c(1,4,3,6))
}
