% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{sim_1_generator}
\alias{sim_1_generator}
\title{This function generates two samples for the simulation of type 1
as described in my master's thesis.}
\usage{
sim_1_generator(n_obs_1, n_obs_2, grid, rho)
}
\arguments{
\item{n_obs_1:}{Number of observations in sample 1}

\item{n_obs_2:}{Number of observations in sample 2}

\item{grid:}{grid at which observations are generated}

\item{rho:}{vector describing the autocorrelation of the process
same length as vector grid}
}
\value{
: list containing the two samples in a functional format
}
\description{
This function generates two samples for the simulation of type 1
as described in my master's thesis.
}
