% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{sim_4_generator}
\alias{sim_4_generator}
\title{This function generates two samples for the simulation of type 4
as described in my master's thesis.}
\usage{
sim_4_generator(n_basis, n_obs_1, n_obs_2, mean, rho, sigma, sigma_shift, grid)
}
\arguments{
\item{n_basis:}{number of fourier basis functions used}

\item{n_obs_1:}{Number of observations in sample 1}

\item{n_obs_2:}{Number of observations in sample 2}

\item{mean:}{vector of same length as grid, that specifies the mean function}

\item{rho:}{vector describing the autocorrelation of the process
same length as vector grid}

\item{sigma:}{vector of same length as grid, that specifies the square root
of the variance function}

\item{sigma_shift:}{vector describing the difference in variance between
the two processes}

\item{grid:}{grid at which observations are generated}
}
\value{
: list containing the two samples in a functional format
}
\description{
This function generates two samples for the simulation of type 4
as described in my master's thesis.
}
