% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_test_alternative.R
\name{tau_realizations_alt}
\alias{tau_realizations_alt}
\title{This function calculates realizations of the cramer-von mises type test
on two given samples for a specified number of permutations}
\usage{
tau_realizations_alt(
  sample1,
  sample2,
  full = FALSE,
  approxQ,
  type = "fourier",
  n_basis,
  grid = NULL,
  domain,
  w_func,
  rho,
  u_sample_func,
  n_func,
  ...
)
}
\arguments{
\item{sample1:}{first sample, specified as a list where each element is one observation}

\item{sample2:}{second sample, specified as a list where each element is one observation}

\item{full:}{Boolean that decides whether all combinations are considered
if FALSE, approxQ has to be provided}

\item{approxQ:}{integer specifying the number of combinations to be used for
the approximation of the critical value}

\item{type:}{type of computation to use (options are: 'fourier', 'eigen', and 'grid' )}

\item{grid:}{grid used for comparison if type == 'grid' is chosen}

\item{domain:}{Domain of the functional observations}

\item{w_func:}{weight function for construction of measure}

\item{rho:}{a vector of length n_basis from a square summable sequence that is used
as scaling terms for the error terms}

\item{u_sample_func:}{function that can be used to sample from for the error terms
around the specified mean}

\item{n_func:}{number of functions used for the Monte-Carlo Integration in the
approximation of tau}

\item{...:}{additional arguments given to u_sample_func}
}
\value{
: Realizations of the t-statistic for the Cramer-von Mises test
}
\description{
This function calculates realizations of the cramer-von mises type test
on two given samples for a specified number of permutations
}
