% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_test_variants.R
\name{tau_test_par}
\alias{tau_test_par}
\title{This function is aimed at the application for my master's thesis and is
experimental. It performs the approximation CvM test parallelized
over the permutations.}
\usage{
tau_test_par(
  cl,
  seeds,
  approxQ = NULL,
  sample1,
  sample2,
  domain = c(0, 1),
  type = "fourier",
  grid = NULL,
  w_func,
  rho,
  u_sample_func,
  n_func,
  ...
)
}
\arguments{
\item{cl:}{cluster object created by parallel package}

\item{seeds:}{Seeds for reproducibility (vector of length approxQ)}

\item{approxQ:}{integer specifying the number of combinations to be used for
the approximation of the critical value}

\item{sample1:}{first sample, specified as a list where each element is one observation}

\item{sample2:}{second sample, specified as a list where each element is one observation}

\item{type:}{type of computation to use (options are: 'fourier', 'eigen', and 'grid' )}

\item{domain:}{Domain of the functional observations}

\item{grid:}{if type grid is chosen, this specifies the grid for the function
comparison}

\item{w_func:}{weight function for construction of measure}

\item{rho:}{a vector of length n_basis from a square summable sequence that is used
as scaling terms for the error terms}

\item{u_sample_func:}{function that can be used to sample from for the error terms
around the specified mean}

\item{n_func:}{number of functions used for the Monte-Carlo Integration in the
approximation of tau}

\item{...:}{additional arguments given to u_sample_func}
}
\value{
Realizations of the t-statistic for the Cramer-von Mises test
}
\description{
This function is aimed at the application for my master's thesis and is
experimental. It performs the approximation CvM test parallelized
over the permutations.
}
