% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_sample_functions.R
\name{u_norm}
\alias{u_norm}
\title{This function is used to sample from multivariate standard normals
and acts as a wrapper around MASS::mvrnorm to interface with the functions
that construct the measure on L2.}
\usage{
u_norm(n_basis)
}
\arguments{
\item{n_basis:}{number of basis functions used in the derivation of
the measure. Specifies the number of error terms that are drawn.}
}
\value{
: A specified number of jointly normal error terms for the construction
of the measure
}
\description{
This function is used to sample from multivariate standard normals
and acts as a wrapper around MASS::mvrnorm to interface with the functions
that construct the measure on L2.
}
