% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_sample_functions.R
\name{u_t}
\alias{u_t}
\title{This function is used to sample from independent t-distributed random variables}
\usage{
u_t(n_basis, df = 2)
}
\arguments{
\item{n_basis:}{number of basis functions used in the derivation of
the measure. Specifies the number of error terms that are drawn.}

\item{df:}{degrees of freedom for the t distribution}
}
\value{
: A specified number of t distributed error terms
}
\description{
This function is used to sample from independent t-distributed random variables
}
