% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_sample_functions.R
\name{u_unif}
\alias{u_unif}
\title{This function is used to sample from independent uniform errors with support [-3,3]
to interface with the functions that construct the measure on L2.}
\usage{
u_unif(n_basis)
}
\arguments{
\item{n_basis:}{number of basis functions used in the derivation of
the measure. Specifies the number of error terms that are drawn.}
}
\value{
: A specified number of independent uniform error terms for the construction
of the measure
}
\description{
This function is used to sample from independent uniform errors with support [-3,3]
to interface with the functions that construct the measure on L2.
}
