% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_construct.R
\name{w_func_construct}
\alias{w_func_construct}
\title{This function implements the method of choosing the weight function
described in my thesis}
\usage{
w_func_construct(sample)
}
\arguments{
\item{sample:}{sample in the list format where each observation contains two
vectors args and vals containing the x and y values}
}
\description{
This function implements the method of choosing the weight function
described in my thesis
}
