% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_construct.R
\name{w_func_construct_2}
\alias{w_func_construct_2}
\title{This function implements method 3 of choosing the weight function
described in my thesis}
\usage{
w_func_construct_2(sample, domain, q)
}
\arguments{
\item{sample:}{functional sample in the fda format}

\item{domain:}{domain of the functional observations}

\item{q:}{chosen quantile for the mean function}
}
\description{
This function implements method 3 of choosing the weight function
described in my thesis
}
