% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_construct.R
\name{w_func_construct_3}
\alias{w_func_construct_3}
\title{This function implements another of choosing the weight function
described in my thesis}
\usage{
w_func_construct_3(sample, domain, q)
}
\arguments{
\item{sample:}{functional sample in the fda format}

\item{domain:}{domain of the functional observations}

\item{q:}{chosen quantile for the mean function}
}
\description{
This function implements another of choosing the weight function
described in my thesis
}
