% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer_von_mises_test.R
\name{cramer_von_mises_tstat}
\alias{cramer_von_mises_tstat}
\title{This function calculates the realization of the cramer-von mises type test
on two given samples}
\usage{
cramer_von_mises_tstat(
  sample1,
  sample2,
  domain = c(0, 1),
  type = "fourier",
  nbasis = NULL,
  grid = NULL,
  mu_measure
)
}
\arguments{
\item{sample2:}{}
}
\description{
This function calculates the realization of the cramer-von mises type test
on two given samples
}
