% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_distribution_function.R
\name{empirical_dist_func}
\alias{empirical_dist_func}
\title{Returns the value of an empirical distribution function of a given sample
for a given function}
\usage{
empirical_dist_func(sample, func, domain = c(0, 1))
}
\arguments{
\item{sample:}{Sample of functional data}

\item{func_b:}{Function to evaluate the empirical distribution function at}

\item{domain:}{Domain of the functions in both the sample and the comparison
function}
}
\value{
An approximated value of the empirical distribution function
}
\description{
Returns the value of an empirical distribution function of a given sample
for a given function
}
