% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_construct.R
\name{fourier_coef_sample}
\alias{fourier_coef_sample}
\title{This function samples a desired distribution of fourier coefficients
around a mean that's determined by the function w}
\usage{
fourier_coef_sample(
  w_func,
  basis = "fourier",
  n_basis,
  rho,
  u_sample_func,
  domain = c(0, 1),
  ...
)
}
\arguments{
\item{w_func:}{function that is large in the areas of the domain the
investigator expects differences in the distribution functions}

\item{basis:}{type of functional basis to use (has to be an orthogonal basis)}

\item{n_basis:}{number of basis functions}

\item{rho:}{a vector of length n_basis from a square summable sequence that is used
as scaling terms for the error terms}

\item{domain:}{vector of two points (start and endpoint of the closed interval)}

\item{u_sample_func:}{function that can be used to sample from for the error terms
around the specified mean}

\item{...:}{further parameters that are given to u_sample_func}
}
\value{
: A vector of length n_basis of sampled fourier coefficients
}
\description{
This function samples a desired distribution of fourier coefficients
around a mean that's determined by the function w
}
