% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_comparison.R
\name{func_comparison}
\alias{func_comparison}
\title{This function checks if one function is bigger than another function for all
points in a closed interval}
\usage{
func_comparison(func_a, func_b, domain = c(0, 1), fourier = FALSE, grid = NULL)
}
\arguments{
\item{func_a:}{Function one}

\item{func_b:}{Function two}

\item{domain:}{interval to check}

\item{fourier:}{If true: func_a and func_b are interpreted as vectors of
fourier coefficients for the same fourier basis. This can be used to solve this
problem more efficiently.}

\item{grid:}{if fourier_coef == FALSE, this decides the grid that is used to check
if one function is always bigger than the other}
}
\value{
TRUE or FALSE depending on wheter func_a is always bigger than
func_b
}
\description{
This function checks if one function is bigger than another function for all
points in a closed interval
}
