% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{func_sample_transform}
\alias{func_sample_transform}
\title{This function transforms a sample of functional data as it is typically generated
by functions from the package fda into the format that is used for example
for the function empirical_dist_func}
\usage{
func_sample_transform(orgnl_sample)
}
\arguments{
\item{orgnl_sample:}{fda object representing multiple observations;
one observation per column of the coef object}
}
\value{
: sample in the format that is used by this package;
list containing elements, each of which is a single observation
}
\description{
This function transforms a sample of functional data as it is typically generated
by functions from the package fda into the format that is used for example
for the function empirical_dist_func
}
