% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_function_generator.R
\name{function_sample}
\alias{function_sample}
\title{This function uses fourier coefficients generated by the function fourier_coef_sample
to generate realizations of the random function associated with the
measure mu.}
\usage{
function_sample(
  w_func,
  basis = "fourier",
  n_basis,
  u_sample_func,
  domain = c(0, 1),
  ...
)
}
\arguments{
\item{w_func:}{function that is large in the areas of the domain the
investigator expects differences in the distribution functions}

\item{basis:}{type of functional basis to use (has to be an orthogonal basis)}

\item{n_basis:}{number of basis functions}

\item{domain:}{vector of two points (start and endpoint of the closed interval)}

\item{u_sample_func:}{function that can be used to sample from for the error terms
around the specified mean}

\item{...:}{further parameters that are given to u_sample_func}
}
\value{
: A functional object that represents the sampled function
}
\description{
This function uses fourier coefficients generated by the function fourier_coef_sample
to generate realizations of the random function associated with the
measure mu.
}
