% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_estimators.R
\name{linear_interpolation}
\alias{linear_interpolation}
\title{This is a helper function that performs linear interpolation for the
linear_mean_estimator function.}
\usage{
linear_interpolation(observation, grid)
}
\arguments{
\item{observation:}{Single Functional Observation}

\item{grid:}{grid to be used for the approximation by linear interpolation}
}
\value{
An approximated observation at the desired grid points.
}
\description{
This is a helper function that performs linear interpolation for the
linear_mean_estimator function.
}
