% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_perm_test_functions.R
\name{perm_test}
\alias{perm_test}
\title{This function performs the task of the permutation test function (called phi
in the corresponding paper)}
\usage{
perm_test(alpha, realizations, realized_value)
}
\arguments{
\item{alpha:}{Significance Level of the test}

\item{realizations:}{vector of realizations of the test statistic over
permutations}

\item{realized_value:}{realized value of the test statistic in original permutation}
}
\value{
: The value of the permutation test function
}
\description{
This function performs the task of the permutation test function (called phi
in the corresponding paper)
}
